refactor(potree2): merge workerPool in Potree2BinParser to fix error when using package in vite project
